+++ tests/test_context.py
@@ -7,7 +7,10 @@ import pyblish.lib
 import pyblish.plugin
 
 from . import lib
-from pyblish.vendor.nose.tools import *
+from pyblish.vendor.nose.tools import (
+    with_setup,
+    raises
+)
 
 
 package_path = pyblish.lib.main_package_path()
