+++ tests/test_plugin.py
@@ -2,6 +2,7 @@ import os
 import shutil
 import tempfile
 
+from pyblish.vendor import mock
 import pyblish.api
 import pyblish.plugin
 from pyblish.vendor.nose.tools import (
@@ -239,118 +240,6 @@ def test_register_host():
     assert "myhost" not in pyblish.plugin.registered_hosts()
 
 
-def test_plugins_from_module():
-    """Getting plug-ins from a module works well"""
-    import types
-
-    module = types.ModuleType("myplugin")
-    code = """
-import pyblish.api
-
-class MyPlugin(pyblish.api.Plugin):
-    def process(self, context):
-        pass
-
-class NotSubclassed(object):
-    def process(self, context):
-        pass
-
-def not_a_plugin():
-    pass
-
-
-class InvalidPlugin(pyblish.api.Plugin):
-    families = False
-
-
-class NotCompatible(pyblish.api.Plugin):
-    hosts = ["not_compatible"]
-
-
-class BadRequires(pyblish.api.Plugin):
-    requires = None
-
-
-class BadHosts(pyblish.api.Plugin):
-    hosts = None
-
-
-class BadFamilies(pyblish.api.Plugin):
-    families = None
-
-
-class BadHosts2(pyblish.api.Plugin):
-    hosts = [None]
-
-
-class BadFamilies2(pyblish.api.Plugin):
-    families = [None]
-
-
-"""
-
-    exec code in module.__dict__
-
-    plugins = pyblish.plugin.plugins_from_module(module)
-
-    assert [p.id for p in plugins] == ["MyPlugin"], plugins
-
-
-@with_setup(lib.setup_empty, lib.teardown)
-def test_discover_globals():
-    """Modules imported in a plug-in are preserved in it's methods"""
-
-    import types
-
-    module = types.ModuleType("myplugin")
-    code = """
-import pyblish.api
-import threading
-
-local_variable_is_present = 5
-
-
-class MyPlugin(pyblish.api.Plugin):
-    def module_is_present(self):
-        return True if threading else False
-
-    def local_variable_is_present(self):
-        return True if local_variable_is_present else False
-
-    def process(self, context):
-        return True if context else False
-
-"""
-
-    exec code in module.__dict__
-    MyPlugin = pyblish.plugin.plugins_from_module(module)[0]
-    assert MyPlugin.id == "MyPlugin"
-
-    assert_true(MyPlugin().process(True))
-    assert_true(MyPlugin().module_is_present())
-    assert_true(MyPlugin().local_variable_is_present())
-
-    try:
-        tempdir = tempfile.mkdtemp()
-        tempplugin = os.path.join(tempdir, "my_plugin.py")
-        with open(tempplugin, "w") as f:
-            f.write(code)
-
-        pyblish.api.register_plugin_path(tempdir)
-        plugins = pyblish.api.discover()
-
-    finally:
-        shutil.rmtree(tempdir)
-
-    assert len(plugins) == 1
-
-    MyPlugin = plugins[0]
-
-    assert_true(MyPlugin().process(True))
-    assert_true(MyPlugin().module_is_present())
-    assert_true(MyPlugin().local_variable_is_present())
-
-
 @with_setup(lib.setup_empty, lib.teardown)
 def test_data_dict():
     """.data is a pure dictionary"""
@@ -464,39 +353,6 @@ def test_plugin_source_path():
 
 
 @with_setup(lib.setup_empty, lib.teardown)
-def test_multi_families():
-    """Instances with multiple families works well"""
-
-    count = {"#": 0}
-
-    class CollectInstance(pyblish.api.Collector):
-        def process(self, context):
-            instance = context.create_instance("MyInstance")
-            instance.data["families"] = ["geometry", "human"]
-
-    class ValidateHumans(pyblish.api.Validator):
-        families = ["human"]
-
-        def process(self, instance):
-            assert "human" in instance.data["families"]
-            count["#"] += 10
-
-    class ValidateGeometry(pyblish.api.Validator):
-        families = ["geometry"]
-
-        def process(self, instance):
-            assert "geometry" in instance.data["families"]
-            count["#"] += 100
-
-    for plugin in (CollectInstance, ValidateHumans, ValidateGeometry):
-        pyblish.api.register_plugin(plugin)
-
-    pyblish.util.publish()
-
-    assert count["#"] == 110, count["#"]
-
-
-@with_setup(lib.setup_empty, lib.teardown)
 def test_register_callback():
     """Callback registration/deregistration works well"""
 
@@ -567,6 +423,31 @@ def test_register_old_plugin():
         requires = "pyblish==0"
 
     pyblish.plugin.register_plugin(MyPlugin)
+
+
+@mock.patch("pyblish.plugin.__explicit_process")
+def test_implicit_explicit_branching(func):
+    """Explicit plug-ins are processed by the appropriate function"""
+
+    # There are two mocks for this (see below); due to
+    # @mock.patch.multiple being a dick.
+
+    class Explicit(pyblish.plugin.ContextPlugin):
+        pass
+
+    pyblish.util.publish(plugins=[Explicit])
+    assert func.call_count == 1, func.call_count
+
+
+@mock.patch("pyblish.plugin.__implicit_process")
+def test_implicit_branching(func):
+    """Implicit plug-ins are processed by the appropriate function"""
+
+    class Implicit(pyblish.plugin.Plugin):
+        pass
+
+    pyblish.util.publish(plugins=[Implicit])
+    assert func.call_count == 1, func.call_count
 
 
 def test_explicit_plugin():
@@ -652,3 +533,19 @@ def test_explicit_action():
         plugin=MyPlugin,
         context=context,
         action="MyAction")
+
+
+def test_explicit_results():
+    """Explicit plug-ins contain results"""
+
+    class Collector(pyblish.plugin.ContextPlugin):
+        order = pyblish.plugin.CollectorOrder
+
+        def process(self, context):
+            self.log.info("logged")
+
+    context = pyblish.util.publish(plugins=[Collector])
+    assert "results" in context.data
+
+    result = context.data["results"][0]
+    assert result["records"][0].msg == "logged"
